Aggiunto metodo inserire esercizio

+++ Mockup-V2/Backend/src/main/java/it/colletta/model/SolutionModel.java
@@ -21,10 +21,10 @@ public class SolutionModel{
     private String id;
     private String solutionText;
     private Date dateSolution;
-    private Double affidability;
+    private int affidability;
     private String authorId;
     public SolutionModel() {
         dateSolution = Calendar.getInstance().getTime();
-        affidability = 0.0;
+        affidability = 0;
     }
 }
